adapt (#28)
