+++ Cargo.toml
@@ -1,0 +1,11 @@
+[workspace.package]
+name = "twimi"
+version = "0.1.0"
+edition = "2021"
+rust-version = "1.65.0"
+
+[workspace]
+members = [
+    "twimi-app",
+    "twimi-core",
+]
