--- src/repositories/models/mod.rs
@@ -1,3 +1,0 @@
-mod kaiin;
-
-pub(super) use kaiin::KaiinTable;
