--- src/usecases/mod.rs
@@ -1,5 +1,0 @@
-pub mod health_check;
-pub mod signup;
-
-pub use health_check::HealthCheckUseCase;
-pub use signup::SignUpUseCase;
