+++ twimi-app/src/bin/main.rs
@@ -1,6 +1,6 @@
 #[tokio::main]
 async fn main() -> std::io::Result<()> {
-    let server = twimi::app::startup::ApiServer::build()?;
+    let server = twimi_app::startup::ApiServer::build()?;
     server.run().await?;
     Ok(())
 }
