+++ twimi-app/src/routes/signup.rs
@@ -2,25 +2,32 @@ use actix_web::{http::StatusCode, web::Json, HttpResponse, ResponseError};
 use serde::{Deserialize, Serialize};
 use time::OffsetDateTime;
 
+use twimi_core::usecases::signup::{SignUp, SignUpInput, SignUpOutput, SignUpUseCaseError};
+
 use super::Inject;
-use crate::usecases::signup::{SignUp, SignUpError, SignUpInput, SignUpOutput};
 
 pub async fn signup(
     usecase: Inject<dyn SignUp>,
     json: Json<SignUpRequestJson>,
 ) -> Result<HttpResponse, SignUpError> {
-    usecase
+    Ok(usecase
         .signup(json.into_inner().into())
         .await
-        .map(|user| HttpResponse::Ok().json(SignUpResponseJson::from(&user)))
+        .map(|v| HttpResponse::Ok().json(SignUpResponseJson::from(&v)))?)
     // .map(|user| HttpResponse::Ok().json(user.as_ref()))
 }
 
+#[derive(Debug, thiserror::Error)]
+#[error(transparent)]
+pub struct SignUpError(#[from] SignUpUseCaseError);
+
 impl ResponseError for SignUpError {
     fn status_code(&self) -> StatusCode {
-        match self {
-            Self::DatabaseError(_) | Self::UnexpectedError(_) => StatusCode::INTERNAL_SERVER_ERROR,
-            Self::ValidationError(_) => StatusCode::BAD_REQUEST,
+        match self.0 {
+            SignUpUseCaseError::DatabaseError(_) | SignUpUseCaseError::UnexpectedError(_) => {
+                StatusCode::INTERNAL_SERVER_ERROR
+            }
+            SignUpUseCaseError::ValidationError(_) => StatusCode::BAD_REQUEST,
         }
     }
 }
