+++ twimi-core/Cargo.toml
@@ -1,8 +1,26 @@
 [package]
 name = "twimi-core"
-version = "0.1.0"
-edition = "2021"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+path = "src/lib.rs"
 
 [dependencies]
+anyhow = "1.0.66"
+argon2 = { version = "0.4.1", features = ["std"] }
+async-trait = "0.1.58"
+rand = "0.8.5"
+shaku = "0.6.1"
+sqlx = { version = "0.6.2", default-features = false, features = [
+    "runtime-actix-rustls",
+    "mysql",
+    "macros",
+    "offline",
+    "time",
+] }
+thiserror = "1.0.37"
+time = { version = "0.3.17", features = ["serde"] }
+unicode-segmentation = "1.10.0"
+validator = "0.16.0"
