+++ twimi-core/src/domain/models/user.rs
@@ -1,43 +1,31 @@
-mod user;
-pub use user::{NewUser, User};
+use time::OffsetDateTime;
 
-use std::marker::PhantomData;
-
-use argon2::{
-    password_hash::{self, SaltString},
-    Algorithm, Argon2, Params, PasswordHasher, Version,
-};
+use super::{ComputeHashError, Id, ValidationError};
+use argon2::{password_hash::SaltString, Algorithm, Argon2, Params, PasswordHasher, Version};
 use unicode_segmentation::UnicodeSegmentation;
 use validator::validate_email;
 
-#[derive(Debug, thiserror::Error)]
-#[error("Failed to validate: {0}")]
-pub struct ValidationError(String);
-
-#[derive(Debug, thiserror::Error)]
-#[error(transparent)]
-pub struct ComputeHashError(#[from] password_hash::Error); // requires "std" feature
-
-pub struct Id<T> {
-    pub value: u64,
-    _marker: PhantomData<T>,
+pub struct NewUser {
+    pub username: Username,
+    pub email: Email,
+    pub password: Password,
 }
 
-impl<T> From<u64> for Id<T> {
-    fn from(value: u64) -> Self {
-        Self {
-            value,
-            _marker: PhantomData,
-        }
+pub struct User {
+    pub id: Id<Self>,
+    pub username: Username,
+    pub email: Email,
+    pub created_at: OffsetDateTime,
+    pub updated_at: OffsetDateTime,
+}
+
+impl User {
+    pub fn id(&self) -> u64 {
+        self.id.value
     }
-}
-
-impl<T> Default for Id<T> {
-    fn default() -> Self {
-        Self {
-            value: Default::default(),
-            _marker: PhantomData,
-        }
+    pub fn set_id(mut self, id: u64) -> Self {
+        self.id = id.into();
+        self
     }
 }
