+++ twimi-core/src/domain/repositories/health_check.rs
@@ -1,0 +1,7 @@
+use async_trait::async_trait;
+use shaku::Interface;
+
+#[async_trait]
+pub trait HealthCheckRepository: Interface {
+    async fn health_check(&self) -> Result<(), sqlx::Error>;
+}
