+++ twimi-core/src/usecases/health_check.rs
@@ -3,17 +3,17 @@ use std::sync::Arc;
 use async_trait::async_trait;
 use shaku::{Component, Interface};
 
-use crate::repositories::HealthCheckRepository;
+use crate::domain::repositories::health_check::HealthCheckRepository;
 
 #[derive(thiserror::Error, Debug)]
-pub enum HealthCheckError {
+pub enum HealthCheckUseCaseError {
     #[error("Failed to acquire a connection from the pool")]
     DatabaseError(#[from] sqlx::Error),
 }
 
 #[async_trait]
 pub trait HealthCheck: Interface {
-    async fn health_check(&self) -> Result<(), HealthCheckError>;
+    async fn health_check(&self) -> Result<(), HealthCheckUseCaseError>;
 }
 
 #[derive(Component)]
@@ -25,7 +25,7 @@ pub struct HealthCheckUseCase {
 
 #[async_trait]
 impl HealthCheck for HealthCheckUseCase {
-    async fn health_check(&self) -> Result<(), HealthCheckError> {
+    async fn health_check(&self) -> Result<(), HealthCheckUseCaseError> {
         self.repository.health_check().await?;
         Ok(())
     }
